# Improve json sql functions by serializing json element into column's buffer direclty