--- a/src/Common/JSONParsers/SimdJSONParser.h
+++ b/src/Common/JSONParsers/SimdJSONParser.h
@@ -8,6 +8,8 @@
 #    include <base/defines.h>
 #    include <simdjson.h>
 #    include "ElementTypes.h"
+#    include <Common/PODArray_fwd.h>
+#    include <Common/PODArray.h>
 
 namespace DB
 {
@@ -15,6 +17,251 @@ namespace ErrorCodes
 {
     extern const int CANNOT_ALLOCATE_MEMORY;
 }
+
+class SimdJSONStringFormatter
+{
+public:
+    explicit SimdJSONStringFormatter(PaddedPODArray<UInt8> & buffer_) : buffer(buffer_) {}
+    inline void comma() { oneChar(','); }
+    /** Start an array, prints [ **/
+    inline void start_array() { oneChar('['); }
+    /** End an array, prints ] **/
+    inline void end_array() { oneChar(']'); }
+    /** Start an array, prints { **/
+    inline void start_object() { oneChar('{'); }
+    /** Start an array, prints } **/
+    inline void end_object() { oneChar('}'); }
+    /** Prints a true **/
+    inline void true_atom()
+    {
+        const char * s = "true";
+        buffer.insert(s, s + 4);
+    }
+    /** Prints a false **/
+    inline void false_atom()
+    {
+        const char * s = "false";
+        buffer.insert(s, s + 5);
+    }
+    /** Prints a null **/
+    inline void null_atom()
+    {
+        const char * s = "null";
+        buffer.insert(s, s + 4);
+    }
+    /** Prints a number **/
+    inline void number(int64_t x)
+    {
+        char number_buffer[24];
+        char * newp = simdjson::fast_itoa(number_buffer, x);
+        buffer.insert(number_buffer, newp);
+    }
+    /** Prints a number **/
+    inline void number(uint64_t x)
+    {
+        char number_buffer[24];
+        char *newp = simdjson::fast_itoa(number_buffer, x);
+        buffer.insert(number_buffer, newp);
+    }
+    /** Prints a number **/
+    inline void number(double x)
+    {
+        char number_buffer[24];
+        // Currently, passing the nullptr to the second argument is
+        // safe because our implementation does not check the second
+        // argument.
+        char * newp = simdjson::internal::to_chars(number_buffer, nullptr, x);
+        buffer.insert(number_buffer, newp);
+    }
+    /** Prints a key (string + colon) **/
+    inline void key(std::string_view unescaped)
+    {
+        string(unescaped);
+        oneChar(':');
+    }
+    /** Prints a string. The string is escaped as needed. **/
+    inline void string(std::string_view unescaped)
+    {
+        oneChar('\"');
+        size_t i = 0;
+        // Fast path for the case where we have no control character, no ", and no backslash.
+        // This should include most keys.
+        //
+        // We would like to use 'bool' but some compilers take offense to bitwise operation
+        // with bool types.
+        constexpr static char needs_escaping[] = {
+            1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0,
+            0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+            0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+            0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+            0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+            0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0};
+        for (; i + 8 <= unescaped.length(); i += 8)
+        {
+            // Poor's man vectorization. This could get much faster if we used SIMD.
+            //
+            // It is not the case that replacing '|' with '||' would be neutral performance-wise.
+            if (needs_escaping[uint8_t(unescaped[i])] | needs_escaping[uint8_t(unescaped[i + 1])]
+                | needs_escaping[uint8_t(unescaped[i + 2])] | needs_escaping[uint8_t(unescaped[i + 3])]
+                | needs_escaping[uint8_t(unescaped[i + 4])] | needs_escaping[uint8_t(unescaped[i + 5])]
+                | needs_escaping[uint8_t(unescaped[i + 6])] | needs_escaping[uint8_t(unescaped[i + 7])])
+            {
+                break;
+            }
+        }
+        for (; i < unescaped.length(); i++)
+        {
+            if (needs_escaping[uint8_t(unescaped[i])])
+            {
+                break;
+            }
+        }
+        // The following is also possible and omits a 256-byte table, but it is slower:
+        // for (; (i < unescaped.length()) && (uint8_t(unescaped[i]) > 0x1F)
+        //      && (unescaped[i] != '\"') && (unescaped[i] != '\\'); i++) {}
+
+        // At least for long strings, the following should be fast. We could
+        // do better by integrating the checks and the insertion.
+        buffer.insert(unescaped.data(), unescaped.data() + i);
+        // We caught a control character if we enter this loop (slow).
+        // Note that we are do not restart from the beginning, but rather we continue
+        // from the point where we encountered something that requires escaping.
+        for (; i < unescaped.length(); i++)
+        {
+            switch (unescaped[i])
+            {
+                case '\"': {
+                    const char * s = "\\\"";
+                    buffer.insert(s, s + 2);
+                }
+                break;
+                case '\\': {
+                    const char * s = "\\\\";
+                    buffer.insert(s, s + 2);
+                }
+                break;
+                default:
+                    if (uint8_t(unescaped[i]) <= 0x1F)
+                    {
+                        // If packed, this uses 8 * 32 bytes.
+                        // Note that we expect most compilers to embed this code in the data
+                        // section.
+                        constexpr static simdjson::escape_sequence escaped[32] = {
+                            {6, "\\u0000"}, {6, "\\u0001"}, {6, "\\u0002"}, {6, "\\u0003"}, {6, "\\u0004"}, {6, "\\u0005"}, {6, "\\u0006"},
+                            {6, "\\u0007"}, {2, "\\b"},     {2, "\\t"},     {2, "\\n"},     {6, "\\u000b"}, {2, "\\f"},     {2, "\\r"},
+                            {6, "\\u000e"}, {6, "\\u000f"}, {6, "\\u0010"}, {6, "\\u0011"}, {6, "\\u0012"}, {6, "\\u0013"}, {6, "\\u0014"},
+                            {6, "\\u0015"}, {6, "\\u0016"}, {6, "\\u0017"}, {6, "\\u0018"}, {6, "\\u0019"}, {6, "\\u001a"}, {6, "\\u001b"},
+                            {6, "\\u001c"}, {6, "\\u001d"}, {6, "\\u001e"}, {6, "\\u001f"}};
+                        auto u = escaped[uint8_t(unescaped[i])];
+                        buffer.insert(u.string, u.string + u.length);
+                    }
+                    else
+                    {
+                        oneChar(unescaped[i]);
+                    }
+            } // switch
+        } // for
+        oneChar('\"');
+    }
+
+    inline void oneChar(char c)
+    {
+        buffer.push_back(c);
+    }
+private:
+    PaddedPODArray<UInt8> & buffer;
+
+};
+
+class SimdJSONStringSerializer
+{
+public:
+    explicit SimdJSONStringSerializer(PaddedPODArray<UInt8> & buffer_) : format(buffer_) {}
+    /** Append an element to the builder (to be printed) **/
+    inline void append(simdjson::dom::element value)
+    {
+        switch (value.type())
+        {
+            case simdjson::dom::element_type::UINT64: {
+                format.number(value.get_uint64().value_unsafe());
+                break;
+            }
+            case simdjson::dom::element_type::INT64: {
+                format.number(value.get_int64().value_unsafe());
+                break;
+            }
+            case simdjson::dom::element_type::DOUBLE: {
+                format.number(value.get_double().value_unsafe());
+                break;
+            }
+            case simdjson::dom::element_type::STRING: {
+                format.string(value.get_string().value_unsafe());
+                break;
+            }
+            case simdjson::dom::element_type::BOOL: {
+                if (value.get_bool().value_unsafe())
+                    format.true_atom();
+                else
+                    format.false_atom();
+                break;
+            }
+            case simdjson::dom::element_type::NULL_VALUE: {
+                format.null_atom();
+                break;
+            }
+            case simdjson::dom::element_type::ARRAY: {
+                append(value.get_array().value_unsafe());
+                break;
+            }
+            case simdjson::dom::element_type::OBJECT: {
+                append(value.get_object().value_unsafe());
+                break;
+            }
+        }
+    }
+    /** Append an array to the builder (to be printed) **/
+    inline void append(simdjson::dom::array value)
+    {
+        format.start_array();
+        auto iter = value.begin();
+        auto end = value.end();
+        if (iter != end)
+        {
+            append(*iter);
+            for (++iter; iter != end; ++iter)
+            {
+                format.comma();
+                append(*iter);
+            }
+        }
+        format.end_array();
+    }
+
+    inline void append(simdjson::dom::object value)
+    {
+        format.start_object();
+        auto pair = value.begin();
+        auto end = value.end();
+        if (pair != end)
+        {
+            append(*pair);
+            for (++pair; pair != end; ++pair)
+            {
+                format.comma();
+                append(*pair);
+            }
+        }
+        format.end_object();
+    }
+
+    inline void append(simdjson::dom::key_value_pair kv)
+    {
+        format.key(kv.key);
+        append(kv.value);
+    }
+private:
+    SimdJSONStringFormatter format;
+};
 
 /// This class can be used as an argument for the template class FunctionJSON.
 /// It provides ability to parse JSONs using simdjson library.