<!---
A technical comment, you are free to remove or leave it as it is when PR is created
The following categories are used in the next scripts, update them accordingly
utils/changelog/changelog.py
tests/ci/cancel_and_rerun_workflow_lambda/app.py
-->
### Changelog category (leave one):

- Performance Improvement



### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):

Instead of serializing json elements into a `std::stringstream`, this PR try to put the serialization result into `ColumnString` direclty.

test result for
```SQL
insert into t_json_str select concat('{"x":', cast(number as String), ',"y":"helloeee", "z":345, "a":{"k1":1, "k2":2}}') from numbers(10000000);
select JSON_QUERY(x, '$.x') from t_json_str Format Null;
```

| | result |
| --- | --- |
| before | 0 rows in set. Elapsed: 3.981 sec. Processed 10.00 million rows, 678.89 MB (2.51 million rows/s., 170.51 MB/s.)|
| after | 0 rows in set. Elapsed: 1.570 sec. Processed 10.00 million rows, 678.89 MB (6.37 million rows/s., 432.43 MB/s.)|

### Documentation entry for user-facing changes

- [ ] Documentation is written (mandatory for new features)

<!---
Directly edit documentation source files in the "docs" folder with the same pull-request as code changes

or

Add a user-readable short description of the changes that should be added to docs.clickhouse.com below.

At a minimum, the following information should be added (but add more as needed).
- Motivation: Why is this function, table engine, etc. useful to ClickHouse users?

- Parameters: If the feature being added takes arguments, options or is influenced by settings, please list them below with a brief explanation.

- Example use: A query or command.
-->


> Information about CI checks: https://clickhouse.com/docs/en/development/continuous-integration/